refactor common mcp implementation
